functions/requirements.txt
firebase-functions
fastapi
pydantic
python-dotenv
requests
google-adk>=0.3.0
google-cloud-aiplatform>=1.64.0
google-auth>=2.33.0
google-generativeai>=0.6.0

